npu2: Add XTS_BDF_MAP wildcard refcount
Currently PID wildcard is programmed into the NPU once and never cleared
up. This works for the bare metal as MSR does not change while the host
OS is running.

However with the device virtualization, we need to keep track of wildcard
entries use and clear them up before switching a GPU from a host to
a guest or vice versa.

This adds refcount to a NPU2, one counter per wildcard entry. The index
is a short lparid which is allocated in opal_npu_map_lpar() and should be
smaller than NPU2_XTS_BDF_MAP_SIZE.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>