phb4: {in|out}_be64 -> phb4_{read|write}_reg
This reduces the number of direct @regs uses so when we will be calling
ultravisor later on, there will be just a few places to change.

While at this, change @phb to @p in phb4_err_inject_finalize to match
the rest of the file and run ':%s/,\t\+/, /g' (vim).

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>