phb4: Always do ASB access for root complexes
PHB4 Root Complex registers only allow 4bytes length access ("4B"),
via ASB or MMIO. We have to use ASB when PHB is fenced and this is not
a hot path anyway so let's simplify things and always use ASB.

However there is a trick. The tests on HV Indirect Data Register (HVIDR,
64bit) show the following:

1. writing a dummy 64bit value to HVIDR and reading it back shows that
top 32bits are written to the registers;

2. reading 4B of the config space results in HVIDR having 2 identical
copies of the 4B value in top and bottom halves;

3. writing to certain registers (notably PCI command/status, may be
others) only works if both halves of the 64bit value have the same copy of
4B value; otherwise writing has no effect;

4. by comparing output from in_le32() (reads from MMIO) and
phb4_read_reg_asb(), the result from phb4_read_reg_asb() requires
byteswapping (which we already do) and so does phb4_write_reg_asb()
(which we do not do).

This uses top 32bits when reading using ASB and duplicates 32bits in both
halves of 64bit value when writing using ASB.

It is arguably a hw bug or an undisclosed feature.

Tested on POWER9 2.1 (pvr 004e 1201); 2.3 (pvr 004e 1203).

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>